Refactor QR code page to be home page
