--- views/account/generate_qr_code/style.css
@@ -1,30 +1,0 @@
-#qrcode-div {
-  height: 300px;
-  width: 300px;
-}
-
-.form-switch.form-switch-xl {
-  margin-bottom: 2rem;
-}
-
-.form-switch.form-switch-xl .form-check-input {
-  height: 2.5rem;
-  width: calc(4rem + 0.75rem);
-  border-radius: 5rem;
-  float: none;
-}
-
-/* Overriding bootstrap color from blue to yellow */
-
-.form-switch .form-check-input:focus {
-  border-color: rgba(0, 0, 0, 0.25);
-  outline: 0;
-  box-shadow: 0 0 0 0 rgba(0, 0, 0, 0);
-  background-image: url("data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' viewBox='-4 -4 8 8'%3e%3ccircle r='3' fill='rgba(0,0,0,0.25)'/></svg>");
-}
-.form-switch .form-check-input:checked {
-  background-color: #ffc107;
-  border-color: #ffc107;
-  border: none;
-  background-image: url("data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' viewBox='-4 -4 8 8'%3e%3ccircle r='3' fill='rgba(255,255,255,1.0)'/></svg>");
-}
