+++ views/partials/navbar.ejs
@@ -1,6 +1,6 @@
 <header class="mb-auto">
   <nav class="px-3 navbar navbar-expand-lg navbar-dark">
-    <a class="navbar-brand text-warning" href="#">G.O.O.N.</a>
+    <a class="navbar-brand text-warning" href="/">G.O.O.N.</a>
     <button
       class="navbar-toggler"
       type="button"
@@ -25,18 +25,17 @@
           <a class="nav-link" href="/aboutus">About Us </a>
         </li>
 
-          <li class="nav-item active">
-            <a class="nav-link" href="/account/profile">Social Media Platforms/Links </a>
-          </li>
-          <li class="nav-item active">
-            <a class="nav-link" href="/account/generate_qr_code">QR Code </a>
-          </li>
-          <li class="nav-item active">
-            <a class="nav-link" href="/account/settings">Settings </a>
-          </li>
-          <li class="nav-item active">
-            <a class="nav-link" href="/account/logout">Logout </a>
-          </li>
+        <li class="nav-item active">
+          <a class="nav-link" href="/account/profile"
+            >Social Media Platforms/Links
+          </a>
+        </li>
+        <li class="nav-item active">
+          <a class="nav-link" href="/account/settings">Settings </a>
+        </li>
+        <li class="nav-item active">
+          <a class="nav-link" href="/account/logout">Logout </a>
+        </li>
       </ul>
     </div>
   </nav>
